+++ ffi/Cargo.toml
@@ -47,4 +47,4 @@ default-engine = [
 ]
 sync-engine = ["delta_kernel/sync-engine"]
 developer-visibility = []
-test_ffi = []
+test-ffi = []
