+++ ffi/examples/read-table/CMakeLists.txt
@@ -41,22 +41,3 @@ if(PRINT_DATA)
   target_compile_definitions(read_table PUBLIC PRINT_ARROW_DATA)
 endif(PRINT_DATA)
 
-# Configuration for the `test_expr` executable
-add_executable(test_expr test_expr.c)
-target_compile_definitions(test_expr PUBLIC DEFINE_DEFAULT_ENGINE)
-target_include_directories(test_expr PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/../../../target/ffi-headers")
-target_link_directories(test_expr PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/../../../target/debug")
-target_link_libraries(test_expr PUBLIC delta_kernel_ffi)
-target_compile_options(test_expr PUBLIC)
-
-if(MSVC)
-  target_compile_options(test_expr PRIVATE /W4 /WX)
-else()
-  target_compile_options(test_expr PRIVATE -Wall -Wextra -Wpedantic -Werror -Wno-strict-prototypes)
-endif()
-
-# Add the kernel expresion -> engine expression test
-include(CTest)
-set(ExprTestRunner "../../../tests/test_expression_visitor/run_test.sh")
-set(ExprExpectedPath "../../../tests/test_expression_visitor/expected.txt")
-add_test(NAME test_expression_visitor COMMAND ${ExprTestRunner} ${ExprExpectedPath})
