+++ ffi/examples/visit-expression/expression_print.h
@@ -1,0 +1,211 @@
+
+#pragma once
+
+#include "expression.h"
+/*************************************************************
+ * Expression Printing
+ ************************************************************/
+
+void print_n_spaces(int n) {
+  if (n == 0)
+    return;
+  printf("  ");
+  print_n_spaces(n - 1);
+}
+void print_tree(ExpressionItem ref, int depth) {
+  switch (ref.type) {
+    case BinOp: {
+      struct BinOp* op = ref.ref;
+      print_n_spaces(depth);
+      switch (op->op) {
+        case Add: {
+          printf("Add\n");
+          break;
+        }
+        case Minus: {
+          printf("Minus\n");
+          break;
+        };
+        case Divide: {
+          printf("Divide\n");
+          break;
+        };
+        case Multiply: {
+          printf("Multiply\n");
+          break;
+        };
+        case LessThan: {
+          printf("LessThan\n");
+          break;
+        };
+        case LessThanOrEqual: {
+          printf("LessThanOrEqual\n");
+          break;
+        }
+        case GreaterThan: {
+          printf("GreaterThan\n");
+          break;
+        };
+        case GreaterThaneOrEqual: {
+          printf("GreaterThanOrEqual\n");
+          break;
+        };
+        case Equal: {
+          printf("Equal\n");
+          break;
+        };
+        case NotEqual: {
+          printf("NotEqual\n");
+          break;
+        };
+        case In: {
+          printf("In\n");
+          break;
+        };
+        case NotIn: {
+          printf("NotIn\n");
+          break;
+        }; break;
+        case Distinct:
+          printf("Distinct\n");
+          break;
+      }
+
+      ExpressionItem left = op->exprs.list[0];
+      ExpressionItem right = op->exprs.list[1];
+      print_tree(left, depth + 1);
+      print_tree(right, depth + 1);
+      break;
+    }
+    case Variadic: {
+      struct Variadic* var = ref.ref;
+      print_n_spaces(depth);
+      switch (var->op) {
+        case And:
+          printf("And\n");
+          break;
+        case Or:
+          printf("Or\n");
+          break;
+        case StructExpression:
+          printf("StructExpression\n");
+          break;
+      }
+      for (size_t i = 0; i < var->exprs.len; i++) {
+        print_tree(var->exprs.list[i], depth + 1);
+      }
+    } break;
+    case Literal: {
+      struct Literal* lit = ref.ref;
+      print_n_spaces(depth);
+      switch (lit->type) {
+        case Integer:
+          printf("Integer");
+          printf("(%d)\n", lit->value.integer_data);
+          break;
+        case Long:
+          printf("Long");
+          printf("(%lld)\n", (long long)lit->value.long_data);
+          break;
+        case Short:
+          printf("Short");
+          printf("(%hd)\n", lit->value.short_data);
+          break;
+        case Byte:
+          printf("Byte");
+          printf("(%hhd)\n", lit->value.byte_data);
+          break;
+        case Float:
+          printf("Float");
+          printf("(%f)\n", (float)lit->value.float_data);
+          break;
+        case Double:
+          printf("Double");
+          printf("(%f)\n", lit->value.double_data);
+          break;
+        case String: {
+          printf("String(%s)\n", lit->value.string_data);
+          break;
+        }
+        case Boolean:
+          printf("Boolean");
+          printf("(%d)\n", lit->value.boolean_data);
+          break;
+        case Timestamp:
+          printf("Timestamp");
+          printf("(%lld)\n", (long long)lit->value.long_data);
+          break;
+        case TimestampNtz:
+          printf("TimestampNtz");
+          printf("(%lld)\n", (long long)lit->value.long_data);
+          break;
+        case Date:
+          printf("Date");
+          printf("(%d)\n", lit->value.integer_data);
+          break;
+        case Binary: {
+          printf("Binary(");
+          for (size_t i = 0; i < lit->value.binary.len; i++) {
+            printf("%02x", lit->value.binary.buf[i]);
+          }
+          printf(")\n");
+          break;
+        }
+        case Decimal: {
+          struct Decimal* dec = &lit->value.decimal;
+          printf("Decimal(%lld,%lld, %d, %d)\n",
+                 (long long)dec->value[0],
+                 (long long)dec->value[1],
+                 dec->scale,
+                 dec->precision);
+          break;
+        }
+        case Null:
+          printf("Null\n");
+          break;
+        case Struct:
+          printf("Struct\n");
+          struct Struct* struct_data = &lit->value.struct_data;
+          for (size_t i = 0; i < struct_data->values.len; i++) {
+            print_n_spaces(depth + 1);
+
+            // Extract field name from field
+            ExpressionItem item = struct_data->fields.list[i];
+            assert(item.type == Literal);
+            struct Literal* lit = item.ref;
+            assert(lit->type == String);
+
+            printf("Field: %s\n", lit->value.string_data);
+            print_tree(struct_data->values.list[i], depth + 2);
+          }
+          break;
+        case Array:
+          printf("Array\n");
+          struct ArrayData* array = &lit->value.array_data;
+          for (size_t i = 0; i < array->exprs.len; i++) {
+            print_tree(array->exprs.list[i], depth + 1);
+          }
+          break;
+      }
+    } break;
+    case Unary: {
+      print_n_spaces(depth);
+      struct Unary* unary = ref.ref;
+      switch (unary->type) {
+        case Not:
+          printf("Not\n");
+          break;
+        case IsNull:
+          printf("IsNull\n");
+          break;
+      }
+      print_tree(unary->sub_expr.list[0], depth + 1);
+      break;
+    }
+    case Column:
+      print_n_spaces(depth);
+      char* column_name = ref.ref;
+      printf("Column(%s)\n", column_name);
+      break;
+  }
+}
