+++ ffi/examples/visit-expression/visit_expr.c
@@ -1,7 +1,8 @@
+
 #include "delta_kernel_ffi.h"
 #include "expression.h"
-int main()
-{
+#include "expression_print.h"
+int main() {
   SharedExpression* pred = get_testing_kernel_expression();
   ExpressionItemList list = construct_predicate(pred);
   ExpressionItem ref = list.list[0];
