+++ ffi/src/lib.rs
@@ -33,7 +33,7 @@ pub mod engine_funcs;
 pub mod expressions;
 pub mod scan;
 pub mod schema;
-#[cfg(feature = "test_ffi")]
+#[cfg(feature = "test-ffi")]
 pub mod test_ffi;
 
 pub(crate) type NullableCvoid = Option<NonNull<c_void>>;
